Update LICENSE
Copyright (c) 2005-2018 Li Shuhang.

Lisence - This is NOT an entirely open source lisence

The following lisence files are written originally in Chinese, while English 
or other languages versions are currently unavailable. If there is any 
inconsistency or ambiguity between the translated versions and the Chinese 
version, the Chinese version shall prevail.

'Chinese' refer to Simplified Chinese / Putonghua used in People's Republic of 
China.